created the pixel formatter class
