+++ LibPngHelper.h
@@ -4,18 +4,20 @@
 #include <png.h>
 #include <string>
 
+// this is self-instantiating because only one should exist at a time
 class LibPngHelper {
     public:
         png_byte colorType;
         png_byte bitDepth;
         png_structp pngPointer;
         png_infop infoPointer;
-        png_bytep *row_pointers;
-        int number_of_passes;
+        png_bytep *rowPointers;
+        int numberOfPasses;
         int width, height;
 
         LibPngHelper();
         std::string readInputPngFile(char* fileName);
-};
+        bool hasAlpha();
+} myLibPngHelper;
 
 #endif
